+++ tests/test_smiles.py
@@ -5,7 +5,7 @@ from rdkit import Chem
 
 from glyles.converter import convert
 from glyles.glycans.factory.factory import MonomerFactory
-from glyles.glycans.utils import Mode, Config
+from glyles.glycans.utils import Config
 from glyles.grammar.parse import Glycan
 from tests.utils import compare_smiles, catch_output, smiles_samples_simple
 
@@ -31,8 +31,7 @@ class TestSMILES:
     @pytest.mark.parametrize("root_orientation", ["n", "a", "b"])
     @pytest.mark.parametrize("iupac, plain, alpha, beta", smiles_samples_simple)
     def test_smiles_poly(self, iupac, plain, alpha, beta, root_orientation):
-        computed = Glycan(iupac, factory=MonomerFactory(), mode=Mode.RDKIT_MODE,
-                          root_orientation=root_orientation).get_smiles()
+        computed = Glycan(iupac, factory=MonomerFactory(), root_orientation=root_orientation).get_smiles()
 
         if root_orientation == "a":
             smiles = alpha